Move code around packages
