--- cspell.json
@@ -1,19 +1,0 @@
-{
-    "version": "0.2",
-    "ignorePaths": [],
-    "dictionaryDefinitions": [],
-    "dictionaries": [],
-    "words": [
-        "Catmull",
-        "cornerstonejs",
-        "crosshairs",
-        "kitware",
-        "Labelmap",
-        "segmentations",
-        "toolgroup",
-        "Viewports",
-        "wado"
-    ],
-    "ignoreWords": [],
-    "import": []
-}
