+++ packages/adapters/src/adapters/Cornerstone3D/RTStruct/index.ts
@@ -1,8 +1,10 @@
-import { generateContourSetsFromLabelmap } from "./utilities/generateContourSetsFromLabelmap";
+import { utilities } from "@cornerstonejs/tools";
 import {
     generateRTSSFromAnnotations,
     generateRTSSFromSegmentations
 } from "./RTSS";
+
+const { generateContourSetsFromLabelmap } = utilities.rtstruct;
 
 export {
     generateContourSetsFromLabelmap,
