--- packages/adapters/src/adapters/Cornerstone3D/RTStruct/utilities/getRTROIObservationsSequence.js
@@ -1,8 +1,0 @@
-export default function getRTROIObservationsSequence(toolData, index) {
-    return {
-        ObservationNumber: index + 1,
-        ReferencedROINumber: index + 1,
-        RTROIInterpretedType: "Todo: type",
-        ROIInterpreter: "Todo: interpreter"
-    };
-}
