--- packages/adapters/src/adapters/Cornerstone3D/RTStruct/utilities/index.js
@@ -1,27 +1,0 @@
-import contourFinder from "./contourFinder";
-import mergePoints from "./mergePoints";
-import detectContourHoles from "./detectContourHoles";
-import { generateContourSetsFromLabelmap } from "./generateContourSetsFromLabelmap";
-
-import AnnotationToPointData from "./AnnotationToPointData";
-
-import getPatientModule from "./getPatientModule";
-import getReferencedFrameOfReferenceSequence from "./getReferencedFrameOfReferenceSequence";
-import getReferencedSeriesSequence from "./getReferencedSeriesSequence";
-import getRTROIObservationsSequence from "./getRTROIObservationsSequence";
-import getRTSeriesModule from "./getRTSeriesModule";
-import getStructureSetModule from "./getStructureSetModule";
-
-export {
-    contourFinder,
-    mergePoints,
-    detectContourHoles,
-    generateContourSetsFromLabelmap,
-    AnnotationToPointData,
-    getPatientModule,
-    getReferencedFrameOfReferenceSequence,
-    getReferencedSeriesSequence,
-    getRTROIObservationsSequence,
-    getRTSeriesModule,
-    getStructureSetModule
-};
