+++ packages/tools/src/utilities/rtstruct/getRTROIObservationsSequence.js
@@ -1,0 +1,8 @@
+export default function getRTROIObservationsSequence(toolData, index) {
+  return {
+    ObservationNumber: index + 1,
+    ReferencedROINumber: index + 1,
+    RTROIInterpretedType: 'Todo: type',
+    ROIInterpreter: 'Todo: interpreter',
+  };
+}
